Creación de migraciones (4-7)

+++ src/main/java/com/polloshermanos/restaurante/PollosHermanosWeb/Dto/DtoSaveProductType.java
@@ -1,4 +1,6 @@
 package com.polloshermanos.restaurante.PollosHermanosWeb.Dto;
 
-public record DtoSaveProductType(float price, String description,Boolean asset) {
+import java.math.BigDecimal;
+
+public record DtoSaveProductType(BigDecimal price, String description, Boolean asset) {
 }
